TFS_Update
git-svn-id: https://dotnetnuke.svn.codeplex.com/svn@65950 142e9c35-602f-4880-9f37-ec48a24daa6a

+++ Trunk/Library/Services/FileSystem/Wrappers/DirectoryWrapper.cs
@@ -28,7 +28,10 @@ namespace DotNetNuke.Services.FileSystem.Internal
     {
         public void Delete(string path, bool recursive)
         {
-            Directory.Delete(path, recursive);
+            if (Exists(path))
+            {
+                Directory.Delete(path, recursive);
+            }
         }
 
         public bool Exists(string path)
